requirements.txt
aiobotocore==2.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aioitertools==0.12.0
aiosignal==1.3.2
arm_pyart==2.0.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
boto3==1.38.27
botocore==1.37.3
Cartopy==0.24.1
certifi==2025.4.26
cffi==1.17.1
cfgrib==0.9.15.0
cftime==1.6.4.post1
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
cmweather==0.3.2
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
dask==2025.5.1
Django==5.2
django-dotenv==1.4.2
django-environ==0.12.0
django-extensions==4.1
django-picklefield==3.3
django-push-notifications==3.2.1
django-q2==1.8.0
eccodes==2.41.0
findlibs==0.1.1
flexcache==0.3
flexparser==0.4
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.1
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
h5netcdf==1.6.1
h5py==3.13.0
honcho==2.0.0
http_ece==1.2.1
idna==3.10
importlib_metadata==8.7.0
jmespath==1.0.1
kiwisolver==1.4.8
lat_lon_parser==1.3.1
locket==1.0.0
Markdown==3.8
matplotlib==3.10.3
mda-xdrlib==0.2.0
multidict==6.4.3
netCDF4==1.7.2
numpy==2.2.5
open-radar-data==0.5.0
packaging==25.0
pandas==2.2.3
partd==1.4.2
pillow==11.2.1
Pint==0.24.4
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.1
psycopg2-binary==2.9.10
py-vapid==1.9.2
pycparser==2.22
pygrib==2.1.6
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywebpush==2.0.3
PyYAML==6.0.2
requests==2.32.3
s3fs==2025.5.1
s3transfer==0.13.0
scipy==1.15.3
shapely==2.1.0
six==1.17.0
sqlparse==0.5.3
stripe==12.1.0
toolz==1.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wrapt==1.17.2
xarray==2025.4.0
xmltodict==0.14.2
xradar==0.9.0
yarl==1.20.0
zipp==3.22.0
